Implement calculator html and css
Add a header with a title, amain with the calculator and a footer
with my name and github link. Add a display to the calculator with
a current-operation and a past-operation area, and a button grid
with all the buttons needed to operate. Customize all those
elements with a responsive web design (at least for mobile) aiming
for a very simple page.

The html code added form the interface with which we simulate a
calculator. The customization garantees that the html code
corresponds to a real calculator and is also responsive.