Implement display and basic calculator functions
Rename display classes and add an alignment class to the display
elements, and add more specific button classes to the buttons
elements. Change delete button color to yellow, operator buttons
color to blue, add margin, alignment to display elements, increase
font-size for current-display, add scroll overflow to past-display
and refactor the css code. Create global variables for the first
and second number of a operation and for the operator, create
global constants to store all buttons and displays, implement
functions to change the global variables, reset them and change
the displays. Add event listeners for the numbers, operations and
equals buttons and implement an approximation function.

The addition of classes to the html elements help customize and
add functionality to them. The customization of these elements
improve the user interface. The addition of global variables and
constants facilitates the calculator logic implementation and the
update functions add responsiveness to the calculator logic. The
event listeners enables a precise control of how button press
interact with the logic and the approximation function restricts
the computations to 10 decimal places.