Implement remaining buttons functions
Implement exponential and factorial operator buttons, implement
decimal button, implement clear and delete buttons, and refactor
the update variables functions.

Except for the exponential and factorial buttons, every other
button is a request of the project. The exponential and factorial
buttons expand the calculator operations with the factorial only
working for non-negative integers. The clear button resets all
current variables and the display while the delete button just
delete the last input of these. The decimal button allows to input
floating points and don't allow to use more than one for the same
number. Lastly, the refactor separates the update function and
the append function from the number variables.