Implement keyboard, add media queries and refactor
Implement keyboard support, correct a bug with division by zero,
add a limit to the number of characters in the current display,
refactor the js code. Change size units in the css file to rem,
fix display overflow bug and add media queries for other screen
sizes.

The keyboard implementation is a requirement of the project. The
addition of media queries and units change to rem are for the
responsivines of the page. All the remaining changes are bug
fixes and code refactoring.